+++ GraphicsEngine/src/TimeManager.h
@@ -1,9 +1,10 @@
 /*
 *
-*	Star Wars: Astromechs
+*	Graphics Engine
 *	By: Darren Kent
 *
 */
+
 #ifndef _TIME_MANAGER_
 #define _TIME_MANAGER_
 
@@ -15,17 +16,17 @@ class TimeManager{
 		void StartTime();									// Sets the start time of the manager
 		void TimeFrame();									// Calculates each value for that frame
 
-		double GetDeltaTime();								// Returns time since last frame in milliseconds
-		double GetDeltaTimeSeconds();						// Returns time since last frame in seconds
-		double GetFramesPerSecond();						// Returns the frames per second
+		float GetDeltaTime();								// Returns time since last frame in milliseconds
+		float GetDeltaTimeSeconds();						// Returns time since last frame in seconds
+		float GetFramesPerSecond();							// Returns the frames per second
 
 		float GetActiveTime();								// Returns the amount of time since the Time Mgr was started	
 	protected:
 	private:
-		double	mStartTime;									// Time in which the time manager started
+		float	mStartTime;									// Time in which the time manager started
 		int		mFrames, mFrameRefreshCap;					// Frame count and the amount of frames we do before refresing calculations
-		double	mCurrentTime, mLastTime, mDeltaTime;		// The current time, last frame's Time, and the time since last frame
-		double	mFramesPerSecond, mLastKeyFrame;			// The average frames per second, and the last frame we checked fps
+		float	mCurrentTime, mLastTime, mDeltaTime;		// The current time, last frame's Time, and the time since last frame
+		float	mFramesPerSecond, mLastKeyFrame;			// The average frames per second, and the last frame we checked fps
 };
 
 #endif // _TIME_MANAGER_
